observers do something, probably correct, TODO make QMatrix visualization for ROS ndoes

+++ sarsa/src/main/java/org/hanns/rl/discrete/observer/impl/BinaryRewardPerStep.java
@@ -23,7 +23,7 @@ public class BinaryRewardPerStep implements Observer{
 	}
 
 	@Override
-	public float getProsperity() { return rewards/steps; }
+	public float getProsperity() { return (float)rewards/(float)steps; }
 
 	@Override
 	public void softReset(boolean randomize) {
